pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "images_pype_dev"

description = "Image gallery for pype.dev"

readme = "README.md"
requires-python = ">=3.7"
license = "MIT"


authors = [
  { name = "Pypeaday", email = "images@pype.dev" },
]


dependencies = ['markata>=0.5.0.dev10', 'Pillow', "textual~=0.2.0", "python-dotenv"]
dynamic = ["version"]

[tool.hatch.version]
path = "images_pype_dev/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pyinstrument",
]

[tool.hatch.envs.default.scripts]
clean = "markata clean"
build = "markata build"
tui = "markata tui"
serve = "python -m http.server 8000 --directory=markout"
clean-build = ['clean', 'build',]
clean-serve = ['clean', 'build', 'serve']
clean-tui = ['clean', 'build', 'tui']
